projects/train/pyproject.toml
[tool.poetry]
name = "train"
description = "Train DeepClean using Lightning"
version = "0.0.1"
authors = ["Alec Gunny"]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"

# torch libs
torchmetrics = "^0.11"
torchaudio = "^0.11"

# lightning dependencies
lightning = "^2.0"
wandb = "^0.15"
jsonargparse = {version = "^4.17", extras = ["signatures"]}

# local dependencies
ml4gw = {path = "../../toolbox/ml4gw", develop = true}
utils = {path = "../utils", develop = true}

[[tool.poetry.source]]
name = "torch"
url = "https://download.pytorch.org/whl/cu118"
secondary = true
default = false

[tool.poetry.group.dev.dependencies]
pytest = "^7.3"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
